requirements.txt
opencv-python
pillow
easyocr